auth

+++ src/app/admin/categories/[id]/page.tsx
@@ -10,6 +10,7 @@ import Loading from "@/app/_components/Loading";
 import LoadingPopup from "@/app/_components/LoadingPopup";
 import CategoryEditorialName from "@/app/_components/CategoryEditorialName";
 import CategoryExisting from "@/app/_components/CategoryExisting";
+import { useAuth } from "@/app/_hooks/useAuth";
 
 // カテゴリの編集・削除のページ
 const Page: React.FC = () => {
@@ -32,6 +33,8 @@ const Page: React.FC = () => {
 
   // カテゴリ配列 (State)。取得中と取得失敗時は null、既存カテゴリが0個なら []
   const [categories, setCategories] = useState<Category[] | null>(null);
+
+  const { token } = useAuth();
 
   // ウェブAPI (/api/categories) からカテゴリの一覧をフェッチする関数の定義
   const fetchCategories = async () => {
@@ -100,12 +103,17 @@ const Page: React.FC = () => {
     setIsSubmitting(true);
 
     try {
+      if (!token) {
+        window.alert("予期せぬ動作：トークンが取得できません。");
+        return;
+      }
       const requestUrl = `/api/admin/categories/${id}`;
       const res = await fetch(requestUrl, {
         method: "PUT",
         cache: "no-store",
         headers: {
           "Content-Type": "application/json",
+          Authorization: token,
         },
         body: JSON.stringify({ name: newCategoryName }),
       });
@@ -137,10 +145,17 @@ const Page: React.FC = () => {
 
     setIsSubmitting(true);
     try {
+      if (!token) {
+        window.alert("予期せぬ動作：トークンができません。");
+        return;
+      }
       const requestUrl = `/api/admin/categories/${id}`;
       const res = await fetch(requestUrl, {
         method: "DELETE",
         cache: "no-store",
+        headers: {
+          Authorization: token,
+        },
       });
 
       if (!res.ok) {
